Function spawn::init

pub fn init(
    kernel_mmi_ref: Arc<MutexIrqSafe<MemoryManagementInfo>>, 
    apic_id: u8, 
    stack: Stack
) -> Result<BootstrapTaskRef, &'static str>

Initializes tasking for the given AP core, including creating a runqueue for it and creating its initial task bootstrapped from the current execution context for that core.