Struct spawn::BootstrapTaskRef

pub struct BootstrapTaskRef { /* fields omitted */ }

A wrapper around a TaskRef that is for bootstrapped tasks.

See spawn::init() and task::bootstrap_task().

This exists such that a bootstrapped task can be marked as exited and removed when being dropped.