spawn

Trait Which

pub trait Which {
    // Required method
    fn which(cmd: &str) -> Result<Cow<'_, str>, Error>;
}

An implementation of which.

Required Methods

fn which(cmd: &str) -> Result<Cow<'_, str>, Error>

Return the absolute path of a command within the PATH.

Errors

If the path could not be resolved, or does not exist.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".