Crate spawn

Process Spawning supporting asynchronous input/output/error capturing, FD passthrough, SetUID mode dropping, SECCOMP filters, and privileged launching.

Structs

Handle

A handle to a child process created via Spawner::spawn() If input/output/error redirection were setup in the Spawner, you can use their related functions to access them.

Spawner

Spawn a child.

Stream

A handle on a process’ Output or Error streams. The Handle can either be used asynchronously to read content as it is filled by the child, or synchronously by calling read_all, which will wait until the child terminates, then collect all output. For async, you can use read_line, or read for an exact byte count.

Enums

HandleError

Errors related to a ProcessHandle

SpawnError

Errors related to the Spawner.

StreamMode

How to handle the standard input/out/error streams