Crate spawn

This crate offers routines for spawning new tasks and convenient builder patterns for customizing new tasks.

The two functions of interest to create a TaskBuilder are:

new_task_builder(): creates a new task for a known, existing function.

new_application_task_builder(): loads a new application crate and creates a new task for that crate’s entry point (main) function.

Structs

BootstrapTaskRef
A wrapper around a TaskRef that is for bootstrapped tasks.

TaskBuilder
A struct that offers a builder pattern to create and customize new Tasks.

Functions

create_idle_task
Spawns an idle task on the given core if specified, otherwise on the current core. Then, it adds adds the new idle task to that core’s runqueue.

init
Initializes tasking for the given AP core, including creating a runqueue for it and creating its initial task bootstrapped from the current execution context for that core.

new_application_task_builder
Creates a builder for a new application Task.

new_task_builder
Creates a builder for a new Task that starts at the given entry point function func and will be passed the given argument.