Crate spawn

This crate offers routines for spawning new tasks and convenient builder patterns for customizing new tasks.

The two functions of interest to create a TaskBuilder are:

new_task_builder(): creates a new task for a known, existing function.

new_application_task_builder(): loads a new application crate and creates a new task for that crate’s entry point (main) function.

Structs

BootstrapTaskRef

A wrapper around a TaskRef for bootstrapped tasks, which are the tasks that represent the first thread of execution on each CPU when it first boots.

TaskBuilder

A struct that offers a builder pattern to create and customize new Tasks.

Functions

cleanup_bootstrap_tasks

Spawns a dedicated task to cleanup all bootstrap tasks by reaping them, i.e., taking their exit value.

init

Initializes tasking for this CPU, including creating a runqueue for it and creating its initial task bootstrapped from the current execution context.

new_application_task_builder

Creates a builder for a new application Task.

new_task_builder

Creates a builder for a new Task that starts at the given entry point function func and will be passed the given argument.