Function spawn::new_application_task_builder

pub fn new_application_task_builder(
    crate_object_file: Path,
    new_namespace: Option<Arc<CrateNamespace>>
) -> Result<TaskBuilder<fn(_: Vec<String>) -> isize, Vec<String>, isize>, &'static str>

Creates a builder for a new application Task.

The new task will start at the application crate’s entry point main function.

Note that the application crate will be loaded and linked during this function, but the actual new application task will not be spawned until TaskBuilder::spawn() is invoked.

Arguments

crate_object_file: the object file that the application crate will be loaded from.

new_namespace: if provided, the new application task will be spawned within the new CrateNamespace, meaning that the new application crate will be linked against the crates within that new namespace. If not provided, the new Task will be spawned within the same namespace as the current task.