Struct spawn::BootstrapTaskRef

pub struct BootstrapTaskRef { /* private fields */ }

A wrapper around a TaskRef for bootstrapped tasks, which are the tasks that represent the first thread of execution on each CPU when it first boots.

When a bootstrap task has done everything it needs to do, it should invoke BootstrapTaskRef::finish() to indicate that it’s finished, which will then mark itself as exited and remove itself from runqueues.

See init() and [task::bootstrap_task()].

Implementations

impl BootstrapTaskRef

pub fn finish(self)

This function represents the final step of each CPU’s initialization procedure.

This function does the following:

Consumes this bootstrap task such that it can no longer be accessed.

Marks this bootstrap task as exited.

Removes this bootstrap task from this CPU’s runqueue.