Struct spawn::TaskBuilder

pub struct TaskBuilder<F, A, R> { /* fields omitted */ }

A struct that offers a builder pattern to create and customize new Tasks.

Note that the new Task will not actually be created until spawn() is invoked.

To create a TaskBuilder, use these functions:

new_task_builder(): creates a new task for a known, existing function.

new_application_task_builder(): loads a new application crate and creates a new task for that crate’s entry point (main) function.

Implementations

impl<F, A, R> TaskBuilder<F, A, R> where
A: Send + 'static,
R: Send + 'static,
F: FnOnce(A) -> R,

pub fn name(self, name: String) -> TaskBuilder<F, A, R>

Set the String name for the new Task.

pub fn argument(self, argument: A) -> TaskBuilder<F, A, R>

Set the argument that will be passed to the new Task’s entry function.

pub fn pin_on_core(self, core_apic_id: u8) -> TaskBuilder<F, A, R>

Pin the new Task to a specific core.

pub fn block(self) -> TaskBuilder<F, A, R>

Set the new Task’s RunState to be Blocked instead of Runnable when it is first spawned. This allows another task to delay the new task’s execution arbitrarily, e.g., to set up other things for the newly-spawned (but not yet running) task.

Note that the new Task will not be Runnable until it is explicitly set as such.

pub fn spawn(self) -> Result<TaskRef, &'static str>

Finishes this TaskBuilder and spawns the new task as described by its builder functions.

This merely makes the new task Runnable, it does not switch to it immediately; that will happen on the next scheduler invocation.

impl<F, A, R> TaskBuilder<F, A, R> where
A: Send + Clone + 'static,
R: Send + 'static,
F: FnOnce(A) -> R + Send + Clone + 'static,

Additional implementation of TaskBuilder to be used for restartable functions. Further restricts the function (F) and argument (A) to implement Clone trait.

pub fn idle(self, core_id: u8) -> TaskBuilder<F, A, R>

Sets this new Task to be the idle task for the given core.

Idle tasks will not be scheduled unless there are no other tasks for the scheduler to choose.

Idle tasks must be restartable, so it is only a possible option when spawning a restartable task. Marking a task as idle is only needed to set up one for each core when that core is initialized, but or to restart an idle task that has exited or failed.

There is no harm spawning multiple idle tasks on each core, but it’s a waste of space.

pub fn spawn_restartable(self) -> Result<TaskRef, &'static str>

Like spawn(), this finishes this TaskBuilder and spawns the new task. It additionally stores the new Task’s function and argument within the Task, enabling it to be restarted upon exit.

This merely makes the new task Runnable, it does not switch to it immediately; that will happen on the next scheduler invocation.